+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Version 1.5.0 (Jun  , 2019)
+## Version 1.5.0 (Sep 04, 2019)
 
 * Added ...
     * added methods expunge, pre_expunge e post_expunge to ApiObject